+++ lib/nigerian_banks_nuban.dart
@@ -1,3 +1,3 @@
-
 export 'src/models/bank.dart';
 export 'src/nigerian_banks_base.dart';
+export 'src/widgets/bank_logo.dart';